Put lib and tests in separate dir and make faster
These tests should be reporting on the structure and correctness of mixer
and its user interface. Therefore the size of the bundles being used is
irrelevant, as long as they are correct definitions and the packages are
real. Trimming down the bundle sizes significantly reduces the time to run
these tests.

Signed-off-by: Tudor Marcu <tudor.marcu@intel.com>